Modify Variable Assignment Placement in EchoClient.java
In this commit. the value for receivedByte is reassigned to input.read() - but outside of the while loop. Once the work is done for writing the output, it will be reset inside the loop. In addition, the print statement at the end of the try block is removed, just in case the tests are checking for additional output.