Add Comments to EchoClient and EchoServer Java Files
This commit adds comments to the EchoClient and EchoServer files to help readers understand the code. To assist with readability, the indentation of the while loops has been set in line with the rest of the code. Previously, they had been indented in further to assist with the writing of the code.